Move TDS2's graph traits into the TDS2 package
